Cleaned up and documented the TurnUndeadTomeAbility.java and UncannyDirectionTomeAbility.java files.

+++ src/main/java/frostvein/sampires/remakepire/abilities/tome/StopTheBleedingTomeAbility.java
@@ -35,7 +35,7 @@ public class StopTheBleedingTomeAbility extends TomeAbility {
      * @param plugin the host plugin object.
      */
     public StopTheBleedingTomeAbility(RemakepirePlugin plugin) {
-        super(plugin, "StopTheBleeding", new String[]{"You learn how to mend the wounds of death itself.", "Crouch within " + PROXIMITY_DISTANCE + " blocks of another player for " + (HEALING_DURATION_TICKS / 20 / 60) + " minute", "to heal one heart for them, restoring their vitality."}, plugin.getConfigManager().getTomeStopTheBleedingCooldown());
+        super(plugin, "StopTheBleeding", new String[]{"You learn how to mend the wounds of death itself.", "Crouch within " + (int)PROXIMITY_DISTANCE + " blocks of another player for " + (HEALING_DURATION_TICKS / 20 / 60) + " minute", "to heal one heart for them, restoring their vitality."}, plugin.getConfigManager().getTomeStopTheBleedingCooldown());
     }
 
     protected boolean useAbility(Player player) {
